requirements.txt
Django==1.9.13
django-filter==0.9.2
django-mptt==0.8.5
django-pg-utils==0.1.5
djangorestframework==3.3.3
factory-boy==2.7.0
psycopg2==2.6.2
pytest==2.9.2
pytest-django==2.9.1
pytest-logging
mixer==5.5.7
Markdown==2.6.2
Pillow==2.8.1
psutil==3.1.1
django-js-reverse==0.6.1
django-registration==2.1.2
django-hashedfilenamestorage==1.0.0
djangorestframework-bulk==0.2.1
django-email-extras==0.3.3
https://storage.googleapis.com/le-downloads/kolibri/kolibri-0.6.0a1-py2.py3-none-any.whl
validators
le-utils>=0.1.5
gunicorn==19.6.0
django-mailgun==0.9.1
pressurecooker==0.0.15
metaphone==0.6
porter2stemmer==1.0
django-postmark==0.1.6
ddtrace==0.8.0
jsonfield==2.0.1
newrelic==2.86.3.70
python-resize-image==1.1.11
sphinx==1.6.4
sphinx-autobuild==0.7.1
sphinx_rtd_theme
sphinx-intl